+++ src/fiflo.c
@@ -2,16 +2,16 @@
 
 #include "errors.c"
 #include "hardware.c"
-#include "render.c"
+#include "logic.c"
 
-void programRound(char base_filename[])
+void programRound(char filename[])
 {
 	while(1)
 	{
 		char pressed_key = unixGetch();
 		keyCheck(pressed_key); // Is it ENTER, BACKSPACE...? More specific.
 		cleanFrame(); // Provide "one-window" rendering in a terminal.
-		window(lines_amount, chars_amount, pressed_key, base_filename);
+		window(lines_amount, chars_amount, pressed_key, filename);
 	}
 }
 
@@ -22,10 +22,11 @@ int main(int argc, char* argv[])
 	{
 		argv[1] = "null.asdf";
 	}
-	baseFilenameLenCheck(argv[1]);
+	filenameLenCheck(argv[1]);
 	argcCheck(argc);
 
-	window(lines_amount, chars_amount, '>', argv[1]);
+	chars_amount--;
+	initWindow(lines_amount, chars_amount, argv[1]);
 	programRound(argv[1]);
 	return 0;
 }