+++ src/hardware.c
@@ -1,12 +1,12 @@
-#include "hardware.h"
-
 // Memory && IO operations.
 
-void pointerCheck(void* pointer) // Empty pointer checker.
+#include "hardware.h"
+
+void pointerCheck(void* pointer) // Eg. malloc or FILE*.
 {
 	if(pointer == NULL)
 	{
-		fputs("Pointer error!\n", stderr);
+		fputs("Memory pointer error.\n", stderr);
 		exit(1);
 	}
 }
@@ -22,15 +22,5 @@ char unixGetch(void) // https://stackoverflow.com/questions/12710582/
 	key = getchar();
 	tcsetattr(STDIN_FILENO, TCSANOW, &oldt);
 	return key;
-}
-
-void saveToFile(char base_filename[])
-{
-	FILE* textfile = fopen(base_filename, "rw");
-	pointerCheck(textfile);
-
-	// TODO
-
-	fclose(textfile);
 }