+++ src/ui.h
@@ -12,9 +12,10 @@
 #define INVERT "\033[7m\0"
 
 void cursor(void);
-static int8_t decimalIntLen(int8_t chars);
+int8_t decimalIntLen(int8_t chars);
 void upperBar(void);
-static void dynamicLen(char base_filename[]);
+
+void dynamicLen(char base_filename[]);
 void lowerBar(int8_t lines, int8_t chars, char key, char base_filename[]);
 void help(void);