+++ src/hardware.c
@@ -15,12 +15,12 @@ char unixGetch(void) // https://stackoverflow.com/questions/12710582/
 {
 	char key;
 	struct termios oldt, newt;
-	tcgetattr(STDIN_FILENO, &oldt);
+	tcgetattr(STDIN_FILENO, &oldt); // Get terminal attribiutes.
 	newt = oldt;
 	newt.c_lflag &= ~(ICANON | ECHO);
-	tcsetattr(STDIN_FILENO, TCSANOW, &newt);
+	tcsetattr(STDIN_FILENO, TCSANOW, &newt); // Set terminal attribiutes.
 	key = getchar();
-	tcsetattr(STDIN_FILENO, TCSANOW, &oldt);
+	tcsetattr(STDIN_FILENO, TCSANOW, &oldt); // Set terminal attribiutes.
 	return key;
 }