+++ src/logic.c
@@ -114,6 +114,12 @@ void allocChars(BUFF_T lines, BUFF_T chars, char key)
 
 	for(line_pos = 1; line_pos <= lines; line_pos++) // Y rendering.
 	{
+		if(line_pos == 1)
+		{
+			printf("%i", line_pos);
+		}
+		// TODO: lines numeration.
+
 		for(char_pos = 1; char_pos <= chars; char_pos++) // X rendering.
 		{
 			printf("%c", text[line_pos - 1][char_pos - 1]);
@@ -146,13 +152,9 @@ void window(BUFF_T lines, BUFF_T chars, char key, char filename[])
 	upperBar();
 	allocChars(lines, chars, key);
 
-	if(chars % windowSize('x') >= 0)
+	if(key != BACKSPACE && chars % windowSize('x') == 0)
 	{
 		vertical_filler++;
-	}
-	else if(chars % windowSize('x') <= 0)
-	{
-		vertical_filler--;
 	}
 
 	for(height = lines; height <= windowSize('y') - vertical_filler; height++)