+++ src/ui.h
@@ -14,8 +14,8 @@
 typedef int16_t BUFF_T;
 
 int8_t decimalIntLen(int8_t chars);
-void upperBar(char *base_fn);
-void dynamicLen(char *filename);
+void upperBar(const char *base_fn);
+void dynamicLen(const char *base_fn);
 void lowerBar(BUFF_T lines, BUFF_T chars, char key);
 void help(void);