+++ docs/CHANGELOG.md
@@ -9,10 +9,11 @@ files use diff-like tools or "git log".
 ## Versions:
 ### v0.5.0 (WIP):
 - Multiple lines writing && rendering.
-- File output operations.
+- File input/output operations.
 - Man page added.
 - Upper bar is back.
 - Absolute paths support.
+- OS check.
 
 ### v0.4.3 (June 10, 2018):
 - Info about min term size fix.